+++ sentry/src/main/java/io/sentry/CheckInStatus.java
@@ -1,0 +1,17 @@
+package io.sentry;
+
+import java.util.Locale;
+import org.jetbrains.annotations.ApiStatus;
+import org.jetbrains.annotations.NotNull;
+
+/** Status of a CheckIn */
+@ApiStatus.Experimental
+public enum CheckInStatus {
+  IN_PROGRESS,
+  OK,
+  ERROR;
+
+  public @NotNull String apiName() {
+    return name().toLowerCase(Locale.ROOT);
+  }
+}