+++ sentry/src/main/java/io/sentry/ExternalOptions.java
@@ -8,6 +8,7 @@ import java.util.Set;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.CopyOnWriteArrayList;
 import java.util.concurrent.CopyOnWriteArraySet;
+import org.jetbrains.annotations.ApiStatus;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
@@ -44,6 +45,10 @@ public final class ExternalOptions {
   private @NotNull Set<String> bundleIds = new CopyOnWriteArraySet<>();
   private @Nullable Boolean enabled;
   private @Nullable Boolean enablePrettySerializationOutput;
+
+  private @Nullable List<String> ignoredCheckIns;
+
+  private @Nullable Boolean sendModules;
 
   @SuppressWarnings("unchecked")
   public static @NotNull ExternalOptions from(
@@ -121,6 +126,10 @@ public final class ExternalOptions {
 
     options.setEnablePrettySerializationOutput(
         propertiesProvider.getBooleanProperty("enable-pretty-serialization-output"));
+
+    options.setSendModules(propertiesProvider.getBooleanProperty("send-modules"));
+
+    options.setIgnoredCheckIns(propertiesProvider.getList("ignored-checkins"));
 
     for (final String ignoredExceptionType :
         propertiesProvider.getList("ignored-exceptions-for-type")) {
@@ -370,5 +379,23 @@ public final class ExternalOptions {
   public void setEnablePrettySerializationOutput(
       final @Nullable Boolean enablePrettySerializationOutput) {
     this.enablePrettySerializationOutput = enablePrettySerializationOutput;
+  }
+
+  public @Nullable Boolean isSendModules() {
+    return sendModules;
+  }
+
+  public void setSendModules(final @Nullable Boolean sendModules) {
+    this.sendModules = sendModules;
+  }
+
+  @ApiStatus.Experimental
+  public void setIgnoredCheckIns(final @Nullable List<String> ignoredCheckIns) {
+    this.ignoredCheckIns = ignoredCheckIns;
+  }
+
+  @ApiStatus.Experimental
+  public @Nullable List<String> getIgnoredCheckIns() {
+    return ignoredCheckIns;
   }
 }