+++ sentry/src/main/java/io/sentry/Sentry.java
@@ -361,7 +361,9 @@ public final class Sentry {
     }
 
     final @NotNull IModulesLoader modulesLoader = options.getModulesLoader();
-    if (modulesLoader instanceof NoOpModulesLoader) {
+    if (!options.isSendModules()) {
+      options.setModulesLoader(NoOpModulesLoader.getInstance());
+    } else if (modulesLoader instanceof NoOpModulesLoader) {
       options.setModulesLoader(
           new CompositeModulesLoader(
               Arrays.asList(
@@ -1015,5 +1017,10 @@ public final class Sentry {
    */
   public static @Nullable BaggageHeader getBaggage() {
     return getCurrentHub().getBaggage();
+  }
+
+  @ApiStatus.Experimental
+  public static @NotNull SentryId captureCheckIn(final @NotNull CheckIn checkIn) {
+    return getCurrentHub().captureCheckIn(checkIn);
   }
 }