+++ sentry/src/main/java/io/sentry/SentryOptions.java
@@ -434,6 +434,12 @@ public class SentryOptions {
   /** Whether to format serialized data, e.g. events logged to console in debug mode */
   private boolean enablePrettySerializationOutput = true;
 
+  /** Whether to send modules containing information about versions. */
+  private boolean sendModules = true;
+
+  /** Contains a list of monitor slugs for which check-ins should not be sent. */
+  @ApiStatus.Experimental private @Nullable List<String> ignoredCheckIns = null;
+
   /**
    * Adds an event processor
    *
@@ -2088,12 +2094,51 @@ public class SentryOptions {
   }
 
   /**
+   * Whether to send modules containing information about versions.
+   *
+   * @return true if modules should be sent.
+   */
+  public boolean isSendModules() {
+    return sendModules;
+  }
+
+  /**
    * Whether to format serialized data, e.g. events logged to console in debug mode
    *
    * @param enablePrettySerializationOutput true if output should be pretty printed
    */
   public void setEnablePrettySerializationOutput(boolean enablePrettySerializationOutput) {
     this.enablePrettySerializationOutput = enablePrettySerializationOutput;
+  }
+
+  /**
+   * Whether to send modules containing information about versions.
+   *
+   * @param sendModules true if modules should be sent.
+   */
+  public void setSendModules(boolean sendModules) {
+    this.sendModules = sendModules;
+  }
+
+  @ApiStatus.Experimental
+  public void setIgnoredCheckIns(final @Nullable List<String> ignoredCheckIns) {
+    if (ignoredCheckIns == null) {
+      this.ignoredCheckIns = null;
+    } else {
+      @NotNull final List<String> filteredIgnoredCheckIns = new ArrayList<>();
+      for (String slug : ignoredCheckIns) {
+        if (!slug.isEmpty()) {
+          filteredIgnoredCheckIns.add(slug);
+        }
+      }
+
+      this.ignoredCheckIns = filteredIgnoredCheckIns;
+    }
+  }
+
+  @ApiStatus.Experimental
+  public @Nullable List<String> getIgnoredCheckIns() {
+    return ignoredCheckIns;
   }
 
   /** Returns the current {@link SentryDateProvider} that is used to retrieve the current date. */
@@ -2349,6 +2394,14 @@ public class SentryOptions {
     }
     if (options.isEnablePrettySerializationOutput() != null) {
       setEnablePrettySerializationOutput(options.isEnablePrettySerializationOutput());
+    }
+
+    if (options.isSendModules() != null) {
+      setSendModules(options.isSendModules());
+    }
+    if (options.getIgnoredCheckIns() != null) {
+      final List<String> ignoredCheckIns = new ArrayList<>(options.getIgnoredCheckIns());
+      setIgnoredCheckIns(ignoredCheckIns);
     }
   }